refactor(web): finish the app platform links unification (#1460)
Move the four campaign-tagged store link sets into
apps/web/src/appPlatformLinks/storeLinks.ts so surfaces no longer import
them from each other's screen modules, and narrow the appPlatformLinks
barrel to what callers actually use.

Let AppPlatformMcpOption carry the locale direction itself instead of
patching it back with a review-dialog RTL override, so the text-only MCP
cell reads correctly wherever the grid is embedded.
