+++ apps/web/src/appPlatformLinks/index.ts
@@ -1,15 +1,13 @@
 export {
   buildAppPlatformOptions,
   type AppPlatformKind,
-  type AppPlatformLabels,
-  type AppPlatformOption,
-  type AppPlatformQrTitles,
-  type AppPlatformStoreKind,
   type AppPlatformStoreLinks,
-  type BuildAppPlatformOptionsInput,
 } from "./appPlatformOptions";
 export { AppPlatformLinksGrid } from "./AppPlatformLinksGrid";
-export { AppPlatformMcpOption, publicMcpServerUrl } from "./AppPlatformMcpOption";
-export { AppPlatformQrCode } from "./AppPlatformQrCode";
-export { AppStoreBadge, GooglePlayBadge, WebAppIcon } from "./badges";
-export { resolveClientPlatform, type ClientPlatform } from "./clientPlatform";
+export { resolveClientPlatform } from "./clientPlatform";
+export {
+  catalogImportStoreLinks,
+  friendInviteStoreLinks,
+  shareAppStoreLinks,
+  webReviewMobilePromptStoreLinks,
+} from "./storeLinks";
