+++ apps/web/src/screens/catalog/CatalogImportSuccessPanel.tsx
@@ -2,17 +2,12 @@ import type { ReactElement } from "react";
 import {
   AppPlatformLinksGrid,
   buildAppPlatformOptions,
+  catalogImportStoreLinks,
   resolveClientPlatform,
-  type AppPlatformStoreLinks,
 } from "../../appPlatformLinks";
 import { getAppConfig } from "../../config";
 import { useI18n } from "../../i18n";
 import { reviewRoute } from "../../routes";
-
-export const catalogImportStoreLinks: AppPlatformStoreLinks = {
-  ios: "https://apps.apple.com/app/apple-store/id6760538964?pt=128797295&ct=catalog_import&mt=8",
-  android: "https://play.google.com/store/apps/details?id=com.flashcardsopensourceapp.app&utm_source=flashcards_website&utm_medium=referral&utm_campaign=catalog_import",
-};
 
 type CatalogImportSuccessPanelProps = Readonly<{
   cardCount: number;
