+++ apps/web/src/screens/invite/FriendInviteScreen.tsx
@@ -11,8 +11,8 @@ import {
 import {
   AppPlatformLinksGrid,
   buildAppPlatformOptions,
+  friendInviteStoreLinks,
   resolveClientPlatform,
-  type AppPlatformStoreLinks,
 } from "../../appPlatformLinks";
 import { useAppErrorDialog } from "../../appError/AppErrorContext";
 import { invalidateServerProgress } from "../../appData/progress/invalidation/progressInvalidation";
@@ -26,11 +26,6 @@ import type {
   SessionInfo,
 } from "../../types";
 import { validateFriendInvitationDisplayName } from "./friendInvitationDisplayName";
-
-export const friendInviteStoreLinks: AppPlatformStoreLinks = {
-  ios: "https://apps.apple.com/app/apple-store/id6760538964?pt=128797295&ct=friend_invite&mt=8",
-  android: "https://play.google.com/store/apps/details?id=com.flashcardsopensourceapp.app&utm_source=flashcards_website&utm_medium=referral&utm_campaign=friend_invite",
-};
 
 type InviteLoadState = "loading" | "inactive" | "error" | "signed_out" | "ready" | "success";
